nbgrader_to_canvas/canvas.py
from canvasapi import Canvas
from flask import session

import requests
import json
import time

from . import settings
from . import db

from .models import Users


class Token:

    def __init__(self, flask_session, user):
        self._flask_session = flask_session
        self._user = user
        self.token_refresh_threshold = 60

    def check(self):
        if self.unexpired() and self.valid_api_key() and self.valid_WWW_Authenticate():
            return True
        else:
            return False

        # expiration_date = self._user.expires_in
        # token_ttl = expiration_date - int(time.time())

        # if token_ttl < self.token_refresh_threshold:
        #     return False
        # elif 'api_key' not in self._flask_session:
        #     return False
        #     # TODO: Figure out what to do here
        #     # Need to know which false is being returned
        # return True



    #If token expires within threshold return false
    def unexpired(self):
        expiration_date = self._user.expires_in
        token_ttl = expiration_date - int(time.time())

        if token_ttl < self.token_refresh_threshold:
            return False
        else:
            return True
     
    def valid_api_key(self):
        if('api_key' in self._flask_session):
            return True
        else:
            return False

    def valid_WWW_Authenticate(self, override=True):
        #TODO: figure out how to spoof this properly instead of just overriding the return to test
        auth_header = {'Authorization': 'Bearer ' + self._flask_session['api_key']}
        r = requests.get(
            '{}users/{}'.format(settings.API_URL, self._flask_session['canvas_user_id']),
            headers=auth_header
        )
        if override or ('WWW-Authenticate' not in r.headers and r.status_code == 200):
            return True
        else:
            return False
        
        

    # def refresh(self):
    #     new_session = refresh_access_token()
    #     if new_session['access_token'] and new_session['expiration_date']:
    #         self._flask_session['api_key'] = new_session['access_token']
    #         self._flask_session['expires_in'] = new_session['expiration_date']
    #         return True
    #     else:
    #         return False

    # Combine old refresh function with refresh_access_token

    def valid_json_response(self, json):
        if 'access_token' not in json:
            return False

        if 'expires_in' not in json:
            return False
        return True
        
        
    def refresh(self):
        """
        Use user's refresh token to get a new access token.

        :rtype: bool
        :returns: True if refresh succeeds. False if json response is invalid, 
            or db does not get updatedDictionary with keys 'access_token' and 'expiration_date'.
        """
        payload = {
                'grant_type': 'refresh_token',
                'client_id': settings.oauth2_id,
                'redirect_uri': settings.oauth2_uri,
                'client_secret': settings.oauth2_key,
                'refresh_token': self._user.refresh_key
            }
        response = requests.post(
            settings.BASE_URL + 'login/oauth2/token',
            data=payload
        )
        # spoofed refresh key is invalid, causing it to redirect. TODO: make the spoofed key 'valid'
        print(response.json())

        #TODO: remove first half of if statement. Find better way to mock this without editing this function  
        if not self.valid_json_response(response.json()):
            app.logger.warning((
                'Access token or expires_in not in json. Bad api key or refresh token.\n'
                'URL: {}\n'
                'Status Code: {}\n'
                'Payload: {}\n'
                'Session: {}'
            ).format(response.url, response.status_code, payload, self._flask_session))
            return False            

        api_key = response.json()['access_token']
        app.logger.info(
            'New access token created\n User: {0}'.format(user.user_id)
        )

        current_time = int(time.time())
        new_expiration_date = current_time + response.json()['expires_in']

        # Update expiration date in db
        if not self.update_db_expiration(new_expiration_date):
            return False

        self._flask_session['api_key'] = api_key
        self._flask_session['expires_in'] = new_expiration_date
        return True

    def update_db_expiration(self, new_expiration_date):
        self._user.expires_in = new_expiration_date
        db.session.commit()

        # Confirm that expiration date has been updated
        updated_user = Users.query.filter_by(user_id=int(self._user.user_id)).first()
        if updated_user.expires_in != new_expiration_date:
            readable_expires_in = time.strftime(
                '%a, %d %b %Y %H:%M:%S',
                time.localtime(updated_user.expires_in)
            )
            readable_new_expiration = time.strftime(
                '%a, %d %b %Y %H:%M:%S',
                time.localtime(new_expiration_date)
            )
            app.logger.error((
                'Error in updating user\'s expiration time in the db:\n'
                'session: {}\n'
                'DB expires_in: {}\n'
                'new_expiration_date: {}'
            ).format(session, readable_expires_in, readable_new_expiration))
            return False
        return True
        

    
    def test(self):
        # Have an API key that shouldn't be expired. Test it to be sure.
        auth_header = {'Authorization': 'Bearer ' + self._flask_session['api_key']}
        r = requests.get(
            '{}users/{}'.format(settings.API_URL, self._flask_session['canvas_user_id']),
            headers=auth_header
        )

        # check for WWW-Authenticate
        # https://canvas.instructure.com/doc/api/file.oauth.html
        if 'WWW-Authenticate' not in r.headers and r.status_code == 200:
            return True
        else:
            # Key is bad. First try to get a new one using refresh

            if refresh():
                return True
            else:
                # Refresh didn't work. Reauthenticate.
                app.logger.info('Reauthenticating:\nflask_session: {}'.format(flask_session))
                return False
        pass

class NbgraderCanvas:

    def __init__(self, api_key, api_URL):
        self._api_key = api_key
        self._api_URL = api_URL
    
    def get_canvas(self):
        return Canvas(self._api_URL, self._api_key)

    def update_token(self, user, flask_session=session):
        token = Token(flask_session, user)
        # # Get the expiration date
        # expiration_date = user.expires_in
        # token_refresh_threshold = 60
        # token_ttl = expiration_date - int(time.time())

        # # refresh the key if it will expire within the next minute, just to make sure the key doesn't expire while responding to a request.
        # if token_ttl < token_refresh_threshold or 'api_key' not in flask_session:           

        #     refresh = refresh_access_token(user)

        #     if refresh['access_token'] and refresh['expiration_date']:
        #         # Success! Set the API Key and Expiration Date
        #         flask_session['api_key'] = refresh['access_token']
        #         flask_session['expires_in'] = refresh['expiration_date']
                
        #         return True
        #     else:
        #         return False
        # else:
            
        #     # Have an API key that shouldn't be expired. Test it to be sure.
        #     auth_header = {'Authorization': 'Bearer ' + flask_session['api_key']}
        #     r = requests.get(
        #         '{}users/{}'.format(settings.API_URL, flask_session['canvas_user_id']),
        #         headers=auth_header
        #     )

        #     # check for WWW-Authenticate
        #     # https://canvas.instructure.com/doc/api/file.oauth.html
        #     if 'WWW-Authenticate' not in r.headers and r.status_code == 200:
                
        #         return True
        #     else:
        #         # Key is bad. First try to get a new one using refresh
        #         new_token = refresh_access_token(user)['access_token']

        #         if new_token:
        #             flask_session['api_key'] = new_token
        #             return True
        #         else:
        #             # Refresh didn't work. Reauthenticate.
        #             app.logger.info('Reauthenticating:\nflask_session: {}'.format(flask_session))
        #             return False    


    # initialize a new Canvas object
    # see https://github.com/ucfopen/canvasapi/issues/238
    # /oauthlogin sets flask_session["api_key"] then redirects to /index
    '''
    def get_canvas():
        try:
            user = Users.query.filter_by(user_id=int(flask_session['canvas_user_id'])).first()

            if check_token_freshness(user):
                canvas = Canvas(settings.API_URL, flask_session['api_key'])
                return canvas
        except Exception as ex:
            app.logger.error("An error occurred connecting to Canvas: {}".format(ex))
            raise

    def refresh_access_token(user):
    """
    Use a user's refresh token to get a new access token.

    :param user: The user to get a new access token for.
    :type user: :class:`Users`

    :rtype: dict
    :returns: Dictionary with keys 'access_token' and 'expiration_date'.
        Values will be `None` if refresh fails.
    """
    refresh_token = user.refresh_key

    payload = {
            'grant_type': 'refresh_token',
            'client_id': settings.oauth2_id,
            'redirect_uri': settings.oauth2_uri,
            'client_secret': settings.oauth2_key,
            'refresh_token': refresh_token
        }
    response = requests.post(
        settings.BASE_URL + 'login/oauth2/token',
        data=payload
    )

    if 'access_token' not in response.json():
        app.logger.warning((
            'Access token not in json. Bad api key or refresh token.\n'
            'URL: {}\n'
            'Status Code: {}\n'
            'Payload: {}\n'
            'Session: {}'
        ).format(response.url, response.status_code, payload, session))
        return {
            'access_token': None,
            'expiration_date': None
        }

    api_key = response.json()['access_token']
    app.logger.info(
        'New access token created\n User: {0}'.format(user.user_id)
    )

    if 'expires_in' not in response.json():
        app.logger.warning((
            'expires_in not in json. Bad api key or refresh token.\n'
            'URL: {}\n'
            'Status Code: {}\n'
            'Payload: {}\n'
            'Session: {}'
        ).format(response.url, response.status_code, payload, session))
        return {
            'access_token': None,
            'expiration_date': None
        }

    current_time = int(time.time())
    new_expiration_date = current_time + response.json()['expires_in']

    # Update expiration date in db
    user.expires_in = new_expiration_date
    db.session.commit()

    # Confirm that expiration date has been updated
    updated_user = Users.query.filter_by(user_id=int(user.user_id)).first()
    if updated_user.expires_in != new_expiration_date:
        readable_expires_in = time.strftime(
            '%a, %d %b %Y %H:%M:%S',
            time.localtime(updated_user.expires_in)
        )
        readable_new_expiration = time.strftime(
            '%a, %d %b %Y %H:%M:%S',
            time.localtime(new_expiration_date)
        )
        app.logger.error((
            'Error in updating user\'s expiration time in the db:\n'
            'session: {}\n'
            'DB expires_in: {}\n'
            'new_expiration_date: {}'
        ).format(session, readable_expires_in, readable_new_expiration))
        return {
            'access_token': None,
            'expiration_date': None
        }

    return {
        'access_token': api_key,
        'expiration_date': new_expiration_date
    }
    

    def check_token_freshness(user):
        # Get the expiration date
        expiration_date = user.expires_in
        token_refresh_threshold = 60
        token_ttl = expiration_date - int(time.time())

        # refresh the key if it will expire within the next minute, just to make sure the key doesn't expire while responding to a request.
        app.logger.debug("Token expires in {} seconds, refreshing if less than threshold of {}".format(token_ttl, token_refresh_threshold))
        if token_ttl < token_refresh_threshold or 'api_key' not in flask_session:
            readable_time = time.strftime('%a, %d %b %Y %H:%M:%S', time.localtime(user.expires_in))

            app.logger.info((
                'Expired refresh token or api_key not in flask_session\n'
                'User: {0}\n'
                'Expiration date in db: {1}\n'
                'Readable expires_in: {2}'
            ).format(user.user_id, user.expires_in, readable_time))

            refresh = refresh_access_token(user)

            if refresh['access_token'] and refresh['expiration_date']:
                # Success! Set the API Key and Expiration Date
                flask_session['api_key'] = refresh['access_token']
                flask_session['expires_in'] = refresh['expiration_date']
                
                return True
            else:
                return False
        else:
            
            # Have an API key that shouldn't be expired. Test it to be sure.
            auth_header = {'Authorization': 'Bearer ' + flask_session['api_key']}
            r = requests.get(
                '{}users/{}'.format(settings.API_URL, flask_session['canvas_user_id']),
                headers=auth_header
            )

            # check for WWW-Authenticate
            # https://canvas.instructure.com/doc/api/file.oauth.html
            if 'WWW-Authenticate' not in r.headers and r.status_code == 200:
                
                return True
            else:
                # Key is bad. First try to get a new one using refresh
                new_token = refresh_access_token(user)['access_token']

                if new_token:
                    flask_session['api_key'] = new_token
                    return True
                else:
                    # Refresh didn't work. Reauthenticate.
                    app.logger.info('Reauthenticating:\nflask_session: {}'.format(flask_session))
                    return False
    '''